BUILD: Merge latest from 2.2_development

+++ src/Gulden/rpcgulden.cpp
@@ -2574,10 +2574,10 @@ static UniValue getlastblocks(const JSONRPCRequest& request)
         
     LogPrintf("getlastblocks requested.\n");
     UniValue result(UniValue::VOBJ);    
-    if (chainActive.Tip()->nHeight > numBlocks)
+    if ((uint64_t)chainActive.Tip()->nHeight > numBlocks)
     {
         CBlockIndex* pIndex = chainActive.Tip();
-        for (int i=0;i<numBlocks;++i)
+        for (uint64_t i=0;i<numBlocks;++i)
         {
             result.push_back(Pair(pIndex->GetBlockHashPoW2().ToString(),pIndex->nHeight));
             pIndex = pIndex->pprev;
